pyproject.toml
[project]
name = "wyoming-nemo-asr"
version = "2.5.0"
description = "Wyoming Server for Nemo ASR"
readme = "README.md"
requires-python = ">=3.11.0,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Michael Hansen", email = "mike@rhasspy.org"}
]
keywords = ["rhasspy", "wyoming", "nemo", "asr", "stt"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Text Processing :: Linguistic",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "wyoming>=1.5.3",
    "nemo_toolkit[all]>=2.3.1"
]

[project.urls]
Homepage = "http://github.com/rhasspy/wyoming-faster-whisper"

[tool.setuptools]
platforms = ["any"]
zip-safe  = true
include-package-data = true

[tool.setuptools.packages.find]
include = ["wyoming_faster_whisper"]
exclude = ["tests"]

[project.scripts]
wyoming-nemo-asr = "wyoming_nemo_asr.__main__:run"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
check_untyped_defs = true
disallow_untyped_defs = true

[project.optional-dependencies]
dev = [
    "black>=22.12.0",
    "flake8>=6.0.0",
    "isort>=5.11.3",
    "mypy>=0.991",
    "pylint>=2.15.9",
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
]
